Separate llama-agent core from CLI
