+++ tools/agent/tools/tool-plan.cpp
@@ -1,5 +1,4 @@
 #include "../tool-registry.h"
-#include "console.h"
 
 #include <sstream>
 #include <string>
@@ -55,11 +54,11 @@ static tool_result plan_execute(const json & args, const tool_context & /* ctx *
         out << "\n" << ANSI_DIM << explanation << ANSI_RESET << "\n";
     }
 
-    // Display the rich plan to the user directly
-    console::log("%s", out.str().c_str());
-
-    // Return terse string to the model to avoid encouraging summarization
-    return {true, "Plan updated.", ""};
+    tool_result result;
+    result.success = true;
+    result.output = out.str();
+    result.content = "Plan updated.";
+    return result;
 }
 
 static tool_def plan_tool = {
